pyproject.toml
[project]
name = "exome-sequencing"
version = "0.1.0"
description = "view the github for more information"
authors = [
    {name = "Osman", email = "72527925+rosman83@users.noreply.github.com"},
]
dependencies = [
    "boto3>=1.34.109",
    "toml>=0.10.2",
    "loguru>=0.7.2",
    "pyyaml>=6.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}
packages = [{include = "*", from="src"}]

[tool.pdm]
distribution = false
